New Boot animation
Thanks to Manuel Möllmann, we now have a boot animation that hopefully
users will approve of. Many users said our old one reminded them of the
DROID X boot animation and that it wasn't 'material' enough. This one is
the opposite of that! It screams 'I'm badass' and demands attention!

For more badass artwork, check out Moelle's work on G+
https://plus.google.com/u/0/+ManuelM%C3%B6llmann

- Added different sizes to ensure that the boot animation fits perfectly on
all our supported devices. Thit boot animation is not a 'one size fits all'
type and so this was neeed.

Change-Id: I2244219b052453a81657a71b8a4bc1b823a00000