scripts: Add ccache build up script
The purpose of this script is to run several builds in a row to fill up
ccache. By default, all supported devices will be built five times but
this can be adjusted with the parameters '-d' and '-n' respectively. See
the help menu for more info

    ./build-up-ccache -h

Notes for other ROMS: Porting this to your own ROM is fairly easy. The
following items needs to be adjusted:

* The default VARIANT
* The PREFIX
* The get_devices function, depending on how your vendorsetup.sh is
  customized

If you have any questions, let me know!

Change-Id: Ib4de4e14757e081c2586caa8b2bece3d9f4a9e7e
Signed-off-by: Nathan Chancellor <nathan@dirtyunicorns.com>